# analyzer/Documentation/openapi_export_20251104.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TokenLense API",
    "description": "API for accessing and visualizing model attribution data",
    "version": "1.0.0"
  },
  "paths": {
...
    "/models/": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Models",
        "description": "Get list of available models",
        "operationId": "get_models_models__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/models/{model}/methods": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Methods",
        "description": "Get attribution methods available for a model",
        "operationId": "get_methods_models__model__methods_get",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelMethods"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/models/{model}/methods/{method}/files": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Files",
        "description": "Get attribution files for a model and method",
        "operationId": "get_files_models__model__methods__method__files_get",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model"
            }
          },
          {
            "name": "method",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Method"
            }
          },
          {
            "name": "include_details",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Details"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelMethodFile"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/performance/system": {
      "get": {
        "tags": [
          "performance"
        ],
        "summary": "Get System Info",
        "description": "Get system information",
        "operationId": "get_system_info_performance_system_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/performance/timing": {
      "get": {
        "tags": [
          "performance"
        ],
        "summary": "Get Timing Results",
        "description": "Get timing results from CSV files",
        "operationId": "get_timing_results_performance_timing_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimingResults"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MethodTimingResult": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "average_time": {
            "type": "number",
            "title": "Average Time"
          },
          "min_time": {
            "type": "number",
            "title": "Min Time"
          },
          "max_time": {
            "type": "number",
            "title": "Max Time"
          },
          "success_rate": {
            "type": "number",
            "title": "Success Rate"
          },
          "tokens_per_second": {
            "type": "number",
            "title": "Tokens Per Second"
          }
        },
        "type": "object",
        "required": [
          "model",
          "method",
          "average_time",
          "min_time",
          "max_time",
          "success_rate",
          "tokens_per_second"
        ],
        "title": "MethodTimingResult",
        "description": "Model for method timing result"
      },
      "ModelInfo": {
        "properties": {
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "models"
        ],
        "title": "ModelInfo",
        "description": "Model for model listing response"
      },
      "ModelMethodFile": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files"
          },
          "file_details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Details",
            "description": "Optional details about each file"
          }
        },
        "type": "object",
        "required": [
          "model",
          "method",
          "files"
        ],
        "title": "ModelMethodFile",
        "description": "Model for file listing response"
      },
      "ModelMethods": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model"
          },
          "methods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Methods"
          }
        },
        "type": "object",
        "required": [
          "model",
          "methods"
        ],
        "title": "ModelMethods",
        "description": "Model for method listing response"
      },
      "PromptTimingResult": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "time": {
            "type": "number",
            "title": "Time"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tokens": {
            "type": "integer",
            "title": "Tokens"
          },
          "tokens_per_second": {
            "type": "number",
            "title": "Tokens Per Second"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "model",
          "method",
          "time",
          "success",
          "tokens",
          "tokens_per_second"
        ],
        "title": "PromptTimingResult",
        "description": "Model for prompt timing result"
      },
      "SystemInfo": {
        "properties": {
          "cpu": {
            "additionalProperties": true,
            "type": "object",
            "title": "Cpu",
            "description": "CPU information"
          },
          "memory": {
            "additionalProperties": true,
            "type": "object",
            "title": "Memory",
            "description": "Memory information"
          },
          "gpu": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu",
            "description": "GPU information if available"
          },
          "python_version": {
            "type": "string",
            "title": "Python Version",
            "description": "Python version"
          },
          "os": {
            "additionalProperties": true,
            "type": "object",
            "title": "Os",
            "description": "Operating system information"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "When the information was collected"
          }
        },
        "type": "object",
        "required": [
          "cpu",
          "memory",
          "python_version",
          "os",
          "timestamp"
        ],
        "title": "SystemInfo",
        "description": "System information model"
      },
      "TimingResults": {
        "properties": {
          "method_timing": {
            "items": {
              "$ref": "#/components/schemas/MethodTimingResult"
            },
            "type": "array",
            "title": "Method Timing"
          },
          "prompt_timing": {
            "items": {
              "$ref": "#/components/schemas/PromptTimingResult"
            },
            "type": "array",
            "title": "Prompt Timing"
          }
        },
        "type": "object",
        "required": [
          "method_timing",
          "prompt_timing"
        ],
        "title": "TimingResults",
        "description": "Model for timing results response"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}